Fix RGB image conversion for display
This fixes a long-standing bug in 2.0beta, which caused RGB(A) images to
be displayed as if they were ABGR images.

The bug resulted from using the 2 methods fixed here, which had
always(?) had the bug but were not used for image display in MM 1.4.x.

Note that the bug was only in image display and transfer to ImageJ.
Image files saved by the buggy versions should still be correct.

The two methods fixed here is not used anywhere else that is RGB-aware,
so the change should be safe.