Move json.org from mmstudio to mmcorej
This actually would have always made sense; even more now that it has
been made clear that json.org is kept primarily as a legacy interface to
mmcorej, due to its use by TaggedImage.

Since mmstudio is for now a single module and depends on mmcorej,
json.org will always be available to it.

Note that we need to continue to "vendor" the json.org source code
(rather than depending on any version on Maven Central), or at least
that is the simplest thing to do, because the ancient version we use has
an incompatible interface with the currently available versions.