Run 'swig -java' from Gradle
This has the nice advantage that (on all platforms) Java modules can be
built without having to depend on the C++ build system (whether it is
VC++ or autotools).

In fact, the C++ build system should later be converted to invoke
'../gradlew :MMCoreJ_wrap:mmcorej:runSwig', and we'll finally only have
to maintain one swig invocation (for Java, anyway; Python is simpler
because it only produces one .py file).

The only caveat is that 'swig' must be on 'PATH', but I don't think
that's an unreasonable requirement, even on Windows (especially in this
era of Chocolatey).